+++ src/main/java/com/ercanbeyen/bankingapplication/repository/ExchangeViewRepository.java
@@ -1,7 +1,7 @@
 package com.ercanbeyen.bankingapplication.repository;
 
 import com.ercanbeyen.bankingapplication.constant.enums.Currency;
-import com.ercanbeyen.bankingapplication.entity.ExchangeView;
+import com.ercanbeyen.bankingapplication.view.ExchangeView;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
