+++ src/main/java/com/ercanbeyen/bankingapplication/service/AccountActivityService.java
@@ -2,7 +2,7 @@ package com.ercanbeyen.bankingapplication.service;
 
 import com.ercanbeyen.bankingapplication.dto.AccountActivityDto;
 import com.ercanbeyen.bankingapplication.dto.request.AccountActivityRequest;
-import com.ercanbeyen.bankingapplication.entity.AccountActivityView;
+import com.ercanbeyen.bankingapplication.view.AccountActivityView;
 import com.ercanbeyen.bankingapplication.option.AccountActivityFilteringOptions;
 
 import java.util.List;
