+++ src/main/java/com/ercanbeyen/bankingapplication/view/AccountActivityView.java
@@ -1,4 +1,4 @@
-package com.ercanbeyen.bankingapplication.entity;
+package com.ercanbeyen.bankingapplication.view;
 
 import com.ercanbeyen.bankingapplication.constant.enums.Currency;
 import com.ercanbeyen.bankingapplication.constant.enums.AccountActivityType;
